requirements.txt
Flask
pandas
pymongo
waitress
mysql-connector-python==8.0.29
